fix some bug

+++ project/src/main/java/com/hungle/mbf/learnspring/repository/UserRepo.java
@@ -3,6 +3,6 @@ package com.hungle.mbf.learnspring.repository;
 import com.hungle.mbf.learnspring.entities.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<Users, Long> {
+public interface UserRepo extends JpaRepository<Users, Integer> {
 
 }
